Convert a QFeatures in Sanssouci4DT object

Convert an assay of an object of class QFeatures in a SansSouci4DT object. A SansSouci4DT object inherits from the SansSouci object class.

SansSouci4DT(obj, i = NULL)

Arguments

obj

An object of class QFeatures.

i

An integer(1) index or a character(1) name of the assay which will be converted. If NULL, the last assay will be selected by default.

Value

An object of class SansSouci4DT.

Author

Manon Gaudin

Examples

numdata <- data.frame(
  S1 = c(2, 5, 9, 10, 7, 8, 6, 8, 6, 7),
  S2 = c(7, 8, 5, 5, 7, 9, 9, 4, 9, 1),
  S3 = c(4, 1, 4, 9, 6, 8, 5, 3, 4, 3),
  S4 = c(9, 8, 9, 8, 3, 6, 4, 1, 4, 5),
  names = paste0("Prot", seq_len(10))
)
obj <- QFeatures::readQFeatures(numdata,
  quantCols = seq_len(4),
  fnames = "names",
  name = "datatest"
)
#> Checking arguments.
#> Loading data as a 'SummarizedExperiment' object.
#> Formatting sample annotations (colData).
#> Formatting data as a 'QFeatures' object.
#> Setting assay rownames.

coldata <- data.frame(
  quantCols = c("S1", "S2", "S3", "S4"),
  Condition = c("C1", "C1", "C2", "C2"),
  Bio.Rep = as.character(seq_len(4))
)
rownames(coldata) <- coldata$quantCols
SummarizedExperiment::colData(obj) <- coldata

ss_obj <- SansSouci4DT(obj)
ss_obj
#> 'SansSouci' object:
#> 	Number of hypotheses:	 10 
#> 	Number of observations:	 4 
#> 	2 samples
#>