Extract selection and post hoc bound to QFeature object

Extract selection and post hoc bound form a SansSouci4DT object to either a data.frame a QFeatures. This function can be executed multiple times on a given QFeatures with different thresholds.

export2dapartoolshed(sanssouci_obj, pval_thr, logfc_thr, qfeature_obj = NULL)

# S3 method for class 'SansSouci4DT'
export2dapartoolshed(sanssouci_obj, pval_thr, logfc_thr, qfeature_obj = NULL)

Arguments

sanssouci_obj

An object of class SansSouci4DT which must be calibrated (see sanssouci::fit function).

pval_thr

A numeric(1), threshold on pvalues.

logfc_thr

A numeric(1), threshold on log Fold Change.

qfeature_obj

An object of class QFeature.

Value

If qfeature_obj is NULL, returns a data.frame containing at least 4 columns :

name_prot : the names of the proteins

pval : the pvalues

logFC : the logFoldChange

sel_n : a binary column for each of the n saved selection

Each selection column includes the following attributes:

the number of selected proteins

the thresholds used

the post hoc bounds on the selection.

The data.frame includes the following attributes :

the number of permutation used in sanssouci calibration (B)

the test function.

If qfeature_obj is not NULL, returns a QFeature with the previously described data.frame stored in metadata(qfeature_obj)$posthoc_selection.

Author

Nicolas Enjalbert Courrech

Examples

n <- 20
p <- 32
X <- as.data.frame(matrix(rnorm(n * p), ncol = n))
colnames(X) <- paste0("S", seq_len(n))
X[, "names"] <- paste0("Prot", seq_len(p))
group <- rep(c(1, 0), length.out = n)
obj <- QFeatures::readQFeatures(X,
  quantCols = seq_len(n),
  fnames = "names",
  name = "datatest"
)
#> Checking arguments.
#> Loading data as a 'SummarizedExperiment' object.
#> Formatting sample annotations (colData).
#> Formatting data as a 'QFeatures' object.
#> Setting assay rownames.

coldata <- data.frame(
  quantCols = paste0("S", seq_len(n)),
  Condition = group,
  Bio.Rep = as.character(seq_len(n))
)
rownames(coldata) <- coldata$quantCols
SummarizedExperiment::colData(obj) <- coldata

ss_obj <- SansSouci4DT(obj)
ss_obj <- sanssouci::fit(ss_obj, alpha = 0.5)

res <- export2dapartoolshed(ss_obj, 0.1, 0.5, qfeature_obj = obj)